Fixed: lambertian values now are properly set as values between 0 and 255 from the UI

+++ Controllers/DTOs/ColorDTO.cs
@@ -8,13 +8,13 @@ namespace Controllers.DTOs
 {
     public class ColorDTO
     {
-        public double Red { get; set; }
-        public double Green { get; set; }
-        public double Blue { get; set; }
+        public int Red { get; set; }
+        public int Green { get; set; }
+        public int Blue { get; set; }
 
         public ColorDTO() { }
 
-        public ColorDTO(double red, double green, double blue)
+        public ColorDTO(int red, int green, int blue)
         {
             Red = red;
             Green = green;
